# docs/swagger.yaml
basePath: /
definitions:
...
  dto.CreateUserAssetRequest:
    properties:
      asset_id:
        type: integer
    required:
    - asset_id
    type: object
  dto.CreateUserRequest:
    properties:
      email:
        type: string
      name:
        maxLength: 50
        minLength: 3
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - name
    - password
    type: object
  dto.LoginRequest:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  dto.LoginResponse:
    properties:
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      token:
        type: string
    type: object
...
  model.UserAsset:
    properties:
      asset_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      user_id:
        type: integer
    type: object
  util.ErrorBody:
    properties:
      error: {}
    type: object
info:
  contact: {}
  description: API for managing users and assets in MF Stock Tracker.
  title: MF Stock Tracker API
  version: "1.0"
paths:
...
  /api/user-assets:
    get:
      description: Get all asset assignments for a user with pagination
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: integer
      - description: 'Number of records to return (default: 50, max: 200)'
        in: query
        name: limit
        type: integer
      - description: 'Number of records to skip (default: 0)'
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              items:
                $ref: '#/definitions/model.UserAsset'
              type: array
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ErrorBody'
      security:
      - BearerAuth: []
      summary: List assets for a user
      tags:
      - user-assets
    post:
      consumes:
      - application/json
      description: Link an asset to a user
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: integer
      - description: Asset assignment payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.CreateUserAssetRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.UserAsset'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ErrorBody'
      security:
      - BearerAuth: []
      summary: Assign asset to user
      tags:
      - user-assets
  /api/user-assets/{userAssetId}:
    delete:
      description: Delete a user-asset link by its ID
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: integer
      - description: User Asset ID
        in: path
        name: userAssetId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ErrorBody'
      security:
      - BearerAuth: []
      summary: Remove asset assignment
      tags:
      - user-assets
  /api/users/:
    post:
      consumes:
      - application/json
      description: Create a new user
      parameters:
      - description: Create user payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.CreateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ErrorBody'
      summary: Create user
      tags:
      - users
  /api/users/{userId}:
    delete:
      description: Delete user by ID
      parameters:
      - description: User ID
        format: int64
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ErrorBody'
      security:
      - BearerAuth: []
      summary: Delete user
      tags:
      - users
  /api/users/login:
    post:
      consumes:
      - application/json
      description: Authenticate user with email and password
      parameters:
      - description: Login payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ErrorBody'
      summary: User login
      tags:
      - users
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"